// components/Github/types.ts
export interface IUser extends Record<string, any> {
	login: string
	id: number
	node_id: string
	avatar_url: string
	gravatar_id: string
	url: string
	html_url: string
	followers_url: string
	following_url: string
	gists_url: string
	starred_url: string
	subscriptions_url: string
	organizations_url: string
	repos_url: string
	events_url: string
	received_events_url: string
	type: string
	site_admin: boolean
	name: string
	company: string
	blog: string
	location: string
	email: string
	hireable: boolean
	bio: string
	public_repos: number
	public_gists: number
	followers: number
	following: number
	twitter_username: string
	created_at: Date
	updated_at: Date
}

export type IResponse = {
	total_count: number
	incomplete_results: boolean
	items: IUser[]
}